+++ docs/pages/components/tabs/api/tabs.js
@@ -83,6 +83,13 @@ export default [
                 type: 'Boolean',
                 values: '-',
                 default: 'false'
+            },
+            {
+                name: '<code>visible</code>',
+                description: 'Item is visible',
+                type: 'Boolean',
+                values: '-',
+                default: 'true'
             }
         ]
     }