+++ src/components/taginput/Taginput.vue
@@ -37,7 +37,18 @@
                 @focus="onFocus"
                 @blur="customOnBlur"
                 @keydown.native="keydown"
-                @select="onSelect"/>
+                @select="onSelect">
+                <template
+                    :slot="defaultSlotName"
+                    slot-scope="props">
+                    <slot
+                        :option="props.option"
+                        :index="props.index" />
+                </template>
+                <template :slot="emptySlotName">
+                    <slot name="empty" />
+                </template>
+            </b-autocomplete>
         </div>
 
         <p v-if="maxtags || maxlength" class="help counter">
@@ -95,6 +106,10 @@
                 type: Array,
                 default: () => [13, 188]
             },
+            removeOnKeys: {
+                type: Array,
+                default: () => [8]
+            },
             allowNew: Boolean
         },
         data() {
@@ -121,6 +136,22 @@
 
             valueLength() {
                 return this.newTag.trim().length
+            },
+
+            defaultSlotName() {
+                return this.hasDefaultSlot ? 'default' : 'dontrender'
+            },
+
+            emptySlotName() {
+                return this.hasEmptySlot ? 'empty' : 'dontrender'
+            },
+
+            hasDefaultSlot() {
+                return !!this.$scopedSlots.default
+            },
+
+            hasEmptySlot() {
+                return !!this.$slots.empty
             },
 
             /**
@@ -202,6 +233,9 @@
             },
 
             keydown(event) {
+                if (this.removeOnKeys.indexOf(event.keyCode) !== -1 && !this.newTag.length) {
+                    this.removeLastTag()
+                }
                 // Stop if is to accept select only
                 if (this.autocomplete && !this.allowNew) return